fix(soft-remove): mark as removed when the snap/tag was with ignore-issues
